Add files via upload
Much better locationcache.txt, created by first determining the grid of every place on earth, then sorting for each grid by population descending and then by distance from grid centroid ascending, and keeping only the first result per grid. The upshot is, instead of representing EM12 as "Midolothian, TX" which is closest to the grid centroid, it is now "Dallas, TX" which gives a much better idea of the location of EM12 for most of us.  If there is no place within the grid that has a population in the data set, I fall back to the old method of closest place to the center of the grid.  Now however, many grids in areas with no place names are not populated with closest placename.